disable foreign keys for seeder truncation

+++ database/migrations/2025_09_09_000004_create_skus_table.php
@@ -9,7 +9,7 @@ return new class extends Migration {
     {
         Schema::create('skus', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('model_id');
+            $table->unsignedInteger('model_id');
             $table->string('name');
             $table->timestamps();
             $table->softDeletes();
